Созданы аналоги EnvironmentProvider и ModelDescriptor для JS окружения
